requirements.txt
openai>=1.37.0
chromadb>=0.5.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
